+++ CHANGES.md
@@ -5,7 +5,8 @@
 * Always install OPAM from source on Alpine until upstreaming
   is complete.
 * Register 4.04 as a mainline compiler as well (fixes OPAM2).
-* Add support for Alpine 3.5 and OpenSUSE 42.2.
+* Add support for Alpine 3.5 and OpenSUSE 42.2, and promote
+  the Alpine:latest images to Alpine 3.5.
 * Do not install camlp4 by default in distributions.
 * Refresh `aspcud` remote proxy with url-escaping fixes
   (via @OCamlPro-Henry in ocaml/opam#2809)